evernote-2026-05-01/pyproject.toml
[project]
name = "evernote-notion-migration"
version = "0.1.0"
description = "Migrate Evernote (.enex) notes into a Notion database, preserving the original Evernote GUID and metadata as database columns."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "notion-client>=2.2.1",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "python-dateutil>=2.9.0",
    "python-dotenv>=1.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
